+++ dwara-db/src/main/java/org/ishafoundation/dwaraapi/db/dao/master/VolumeDao.java
@@ -15,6 +15,8 @@ public interface VolumeDao extends CrudRepository<Volume,String> {
 		
 	List<Volume> findAllByType(Volumetype volumetype);
 	
+	List<Volume> findAllByTypeAndImportedIsFalse(Volumetype volumetype);
+	
 	List<Volume> findAllByStoragetypeAndType(Storagetype storagetype, Volumetype volumetype);
 
 	List<Volume> findAllByCopyId(int copyId);